requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5

SQLAlchemy==2.0.25
pymysql==1.1.2
psycopg2-binary==2.9.9
psycopg[binary]>=3.2
cloudinary==1.36.0

gunicorn==21.2.0
Werkzeug==2.3.7

python-dotenv==1.0.1
resend
razorpay==1.4.2
setuptools<81
